Begun work on new sound system, removed SDL_Mixer
Extremely alpha sound support can be enabled by changing sound_enabled to true in quig.cpp. It plays a high/low tone when the A-button is pressed.
also, apparently, a bunch of licenses were not in other-licenses/ for some reason, fixed that

I may not actually end up using Blip_Buffer, but as of currently, it is what I'm expecting to go ahead with.